Feature/user service course service

+++ back-end/apps/course-service/src/main/java/com/coursedemy/course/service/CategoryService.java
@@ -12,5 +12,5 @@ public interface CategoryService {
 
     void createCategory(CategoryDTO categoryDTO);
 
-    void deleteCategory(Integer id);
+    void deleteCategory(Long id);
 }
